Refactor game_state directory for improved organization and maintainability
- Create TileDataFactory to consolidate TileData creation patterns
  - Add specialized factory methods for different tile types
  - Replace 40+ line placeholder creation with single factory call
  - Improve code reusability and consistency

- Extract detection functions into TilePropertyDetector class
  - Move 8 detection functions from enhanced_tile_creator.py to dedicated class
  - Improve organization and single responsibility principle
  - Maintain consistent API while grouping related functionality

- Simplify screen analysis with base filtering function
  - Add _filter_tiles_by() to reduce code duplication
  - Refactor get_walkable_tiles, get_interactive_tiles, etc. to use common base
  - Improve maintainability and consistency across filtering functions

- Update module exports and test compatibility
  - Export new factory and detector classes in __init__.py
  - Update tests to use new TilePropertyDetector API
  - Maintain backward compatibility for existing functionality

All tests pass and code quality checks pass. This refactoring reduces
code duplication by ~300 lines while improving organization and clarity.
