+++ open_llms_play_pokemon/game_state/__init__.py
@@ -1,7 +1,11 @@
 from .game_state import PokemonRedGameState
 from .memory_reader import PokemonRedMemoryReader
+from .tile_data_factory import TileDataFactory
+from .tile_property_detector import TilePropertyDetector
 
 __all__ = [
     "PokemonRedGameState",
     "PokemonRedMemoryReader",
+    "TileDataFactory",
+    "TilePropertyDetector",
 ]
